+++ src/Components/Tech/SingleTech.js
@@ -1,12 +1,23 @@
 import React from "react";
-import { SingleTechCard, SingleTechDiv, SingleTechImg, SingleTechOutOfStock, TechProductName, TechProductPrice } from "./StyledSingleTech";
+import { 
+    AlreadyInCart,
+    AlreadyInCartImg,
+    SingleTechCard, 
+    SingleTechDiv, 
+    SingleTechImg, 
+    SingleTechOutOfStock, 
+    TechProductName, 
+    TechProductPrice 
+} from "./StyledSingleTech";
+import shoppingcartwhite from "../../Images/shoppingcartwhite.png";
 
 class SingleTech extends React.Component {
     constructor(props) {
         super(props);
         this.state = {
             currencyIndex: 0,
-            activeCurrencySymbol: "$"
+            activeCurrencySymbol: "$",
+            alreadyInCart: false
         };
         this.addProductToCart = this.addProductToCart.bind(this);
     }
@@ -25,10 +36,19 @@ class SingleTech extends React.Component {
                 }
             }
         }
+        if (localStorage.getItem("shoppingCart") !== null) {
+            const temporary_cart = JSON.parse(localStorage.getItem("shoppingCart"));
+            for (let i=0; i<temporary_cart.length; i++) {
+                if (temporary_cart[i].id === this.props.singleTech.id) {
+                    this.setState({alreadyInCart: true});
+                    break;
+                }
+            }
+        }
     }
 
     addProductToCart() {
-        if (this.props.singleTech.inStock) {
+        if (this.props.singleTech.inStock && !this.state.alreadyInCart) {
             localStorage.setItem("beforeCart", this.props.singleTech.id);
             window.location.pathname = "/add-to-cart";
         }
@@ -44,6 +64,7 @@ class SingleTech extends React.Component {
                         <div>
                             <SingleTechImg src={this.props.singleTech.gallery}/>
                             {!this.props.singleTech.inStock && <SingleTechOutOfStock>OUT OF STOCK</SingleTechOutOfStock>}
+                            {this.state.alreadyInCart && <AlreadyInCart><AlreadyInCartImg src={shoppingcartwhite}/></AlreadyInCart>}
                         </div>
                         <TechProductName>{this.props.singleTech.name}</TechProductName>
                         <TechProductPrice>{this.state.activeCurrencySymbol}{this.props.singleTech.prices[this.state.currencyIndex].amount}</TechProductPrice>
